делает тз модуляция
